pyproject.toml
[project]
name = "trip_planner"
version = "0.1.0"
description = "Trip Planner using crewAI"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai==0.85.0",
    "crewai-tools==0.14.0",
    "setuptools==80.9.0",
    "arize-phoenix-otel==0.12.1",
    "openinference-instrumentation-crewai==0.1.10",
    "google-search-results==2.4.2"
]

[tool.setuptools.packages.find]
where = ["src"]
